Another step on the refactor
* Separated out some functionality away from both the symbols and
  parsing
* Major clean-up for some of the parsing code
