+++ nodeutil/node_helpers.go
@@ -1,0 +1,23 @@
+package nodeutil
+
+import sitter "github.com/smacker/go-tree-sitter"
+
+// NamedChildrenOf gets all named children of a given node
+func NamedChildrenOf(node *sitter.Node) []*sitter.Node {
+	count := int(node.NamedChildCount())
+	children := make([]*sitter.Node, count)
+	for i := 0; i < count; i++ {
+		children[i] = node.NamedChild(i)
+	}
+	return children
+}
+
+// UnnamedChildrenOf gets all the named + unnamed children of a given node
+func UnnamedChildrenOf(node *sitter.Node) []*sitter.Node {
+	count := int(node.ChildCount())
+	children := make([]*sitter.Node, count)
+	for i := 0; i < count; i++ {
+		children[i] = node.Child(i)
+	}
+	return children
+}
